change elevation type in ElevationPoint to float64

+++ backend/db/types.go
@@ -62,7 +62,7 @@ type OrderedElevationPoint interface {
 
 type ElevationPoint struct {
 	Distance  float64
-	Elevation int
+	Elevation float64
 }
 
 func (ep ElevationPoint) Less(other ElevationPoint) bool {
